fix: api reference switching issue
