+++ packages/ui/app/src/playground/utils/serialize.ts
@@ -1,0 +1,97 @@
+import { assertNever, isNonNullish } from "@fern-ui/core-utils";
+import { compact } from "lodash-es";
+import { ResolvedFormDataRequestProperty, ResolvedHttpRequestBodyShape } from "../../resolver/types";
+import { blobToDataURL } from "../fetch-utils/blobToDataURL";
+import { PlaygroundFormStateBody, ProxyRequest, SerializableFile, SerializableFormDataEntryValue } from "../types";
+
+export async function serializeFormStateBody(
+    environment: string,
+    shape: ResolvedHttpRequestBodyShape | undefined,
+    body: PlaygroundFormStateBody | undefined,
+    usesApplicationJsonInFormDataValue: boolean,
+): Promise<ProxyRequest.SerializableBody | undefined> {
+    if (shape == null || body == null) {
+        return undefined;
+    }
+
+    switch (body.type) {
+        case "json":
+            return { type: "json", value: body.value };
+        case "form-data": {
+            const formDataValue: Record<string, SerializableFormDataEntryValue> = {};
+            for (const [key, value] of Object.entries(body.value)) {
+                switch (value.type) {
+                    case "file":
+                        formDataValue[key] = {
+                            type: "file",
+                            value: await serializeFile(environment, value.value),
+                        };
+                        break;
+                    case "fileArray":
+                        formDataValue[key] = {
+                            type: "fileArray",
+                            value: (
+                                await Promise.all(value.value.map((value) => serializeFile(environment, value)))
+                            ).filter(isNonNullish),
+                        };
+                        break;
+                    case "json": {
+                        if (shape.type !== "formData") {
+                            return undefined;
+                        }
+                        const property = shape.properties.find((p) => p.key === key && p.type === "bodyProperty") as
+                            | ResolvedFormDataRequestProperty.BodyProperty
+                            | undefined;
+
+                        // check if the json value is a string and performa a safe parse operation to check if the json is stringified
+                        if (typeof value.value === "string") {
+                            value.value = safeParse(value.value);
+                        }
+
+                        formDataValue[key] = {
+                            ...value,
+                            // this is a hack to allow the API Playground to send JSON blobs in form data
+                            // revert this once we have a better solution
+                            contentType:
+                                compact(property?.contentType)[0] ??
+                                (usesApplicationJsonInFormDataValue ? "application/json" : undefined),
+                        };
+                        break;
+                    }
+                    default:
+                        assertNever(value);
+                }
+            }
+            return { type: "form-data", value: formDataValue };
+        }
+        case "octet-stream":
+            return { type: "octet-stream", value: await serializeFile(environment, body.value) };
+        default:
+            assertNever(body);
+    }
+}
+
+async function serializeFile(environment: string, file: File | undefined): Promise<SerializableFile | undefined> {
+    if (file == null || !isFile(file)) {
+        return undefined;
+    }
+    return {
+        name: file.name,
+        lastModified: file.lastModified,
+        size: file.size,
+        type: file.type,
+        dataUrl: await blobToDataURL(environment, file),
+    };
+}
+
+function isFile(value: any): value is File {
+    return value instanceof File;
+}
+
+function safeParse(value: string): unknown {
+    try {
+        return JSON.parse(value);
+    } catch {
+        return value;
+    }
+}
