+++ packages/ui/app/src/playground/utils/utils.ts
@@ -1,0 +1,135 @@
+import { isPlainObject, visitDiscriminatedUnion } from "@fern-ui/core-utils";
+import {
+    ResolvedHttpRequestBodyShape,
+    ResolvedTypeDefinition,
+    ResolvedTypeShape,
+    dereferenceObjectProperties,
+    unwrapReference,
+    visitResolvedHttpRequestBodyShape,
+} from "../../resolver/types";
+
+export function castToRecord(value: unknown): Record<string, unknown> {
+    if (!isPlainObject(value)) {
+        return {};
+    }
+    return value;
+}
+
+export function isExpandable(
+    valueShape: ResolvedTypeShape,
+    currentValue: unknown,
+    types: Record<string, ResolvedTypeDefinition>,
+): boolean {
+    return visitDiscriminatedUnion(unwrapReference(valueShape, types), "type")._visit<boolean>({
+        object: () => false,
+        discriminatedUnion: () => false,
+        undiscriminatedUnion: () => false,
+        enum: () => false,
+        optional: (optional) => isExpandable(optional.shape, currentValue, types),
+        list: () => Array.isArray(currentValue) && currentValue.length > 0,
+        set: () => Array.isArray(currentValue) && currentValue.length > 0,
+        map: () => isPlainObject(currentValue) && Object.keys(currentValue).length > 0,
+        unknown: () => false,
+        _other: () => false,
+        primitive: () => false,
+        literal: () => false,
+        alias: (alias) => isExpandable(alias.shape, currentValue, types),
+    });
+}
+
+export function hasRequiredFields(
+    bodyShape: ResolvedHttpRequestBodyShape,
+    types: Record<string, ResolvedTypeDefinition>,
+): boolean {
+    return visitResolvedHttpRequestBodyShape(bodyShape, {
+        formData: (formData) =>
+            formData.properties.some((property) =>
+                visitDiscriminatedUnion(property, "type")._visit<boolean>({
+                    file: (file) => !file.isOptional,
+                    fileArray: (fileArray) => !fileArray.isOptional,
+                    bodyProperty: (bodyProperty) => hasRequiredFields(bodyProperty.valueShape, types),
+                    _other: () => false,
+                }),
+            ) ?? true,
+        bytes: (bytes) => !bytes.isOptional,
+        typeShape: (shape) =>
+            visitDiscriminatedUnion(unwrapReference(shape, types), "type")._visit({
+                primitive: () => true,
+                literal: () => true,
+                object: (object) =>
+                    dereferenceObjectProperties(object, types).some((property) =>
+                        hasRequiredFields(property.valueShape, types),
+                    ),
+                undiscriminatedUnion: () => true,
+                discriminatedUnion: () => true,
+                enum: () => true,
+                optional: () => false,
+                list: () => true,
+                set: () => true,
+                map: () => true,
+                unknown: () => true,
+                alias: (alias) => hasRequiredFields(alias.shape, types),
+                _other: () => true,
+            }),
+    });
+}
+
+export function hasOptionalFields(
+    bodyShape: ResolvedHttpRequestBodyShape,
+    types: Record<string, ResolvedTypeDefinition>,
+): boolean {
+    return visitResolvedHttpRequestBodyShape(bodyShape, {
+        formData: (formData) =>
+            formData.properties.some((property) =>
+                visitDiscriminatedUnion(property, "type")._visit<boolean>({
+                    file: (file) => file.isOptional,
+                    fileArray: (fileArray) => fileArray.isOptional,
+                    bodyProperty: (bodyProperty) => hasOptionalFields(bodyProperty.valueShape, types),
+                    _other: () => false,
+                }),
+            ) ?? false,
+        bytes: (bytes) => bytes.isOptional,
+        typeShape: (shape) =>
+            visitDiscriminatedUnion(unwrapReference(shape, types), "type")._visit({
+                primitive: () => false,
+                literal: () => false,
+                object: (object) =>
+                    dereferenceObjectProperties(object, types).some((property) =>
+                        hasOptionalFields(property.valueShape, types),
+                    ),
+                undiscriminatedUnion: () => false,
+                discriminatedUnion: () => false,
+                enum: () => false,
+                optional: () => true,
+                list: () => false,
+                set: () => false,
+                map: () => false,
+                unknown: () => false,
+                alias: (alias) => hasOptionalFields(alias.shape, types),
+                _other: () => false,
+            }),
+    });
+}
+
+export const ENUM_RADIO_BREAKPOINT = 5;
+
+export function shouldRenderInline(
+    typeReference: ResolvedTypeShape,
+    types: Record<string, ResolvedTypeDefinition>,
+): boolean {
+    return visitDiscriminatedUnion(unwrapReference(typeReference, types), "type")._visit({
+        primitive: () => true,
+        literal: () => true,
+        object: () => false,
+        map: () => false,
+        undiscriminatedUnion: () => false,
+        discriminatedUnion: () => false,
+        enum: (_enum) => true,
+        optional: () => false,
+        list: () => false,
+        set: () => false,
+        unknown: () => false,
+        alias: (alias) => shouldRenderInline(alias.shape, types),
+        _other: () => false,
+    });
+}
